% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-pubg.R
\name{fetch_filtered_pubg_player_stats}
\alias{fetch_filtered_pubg_player_stats}
\title{Fetch filtered stats for PUBG Player}
\usage{
fetch_filtered_pubg_player_stats(pubgNickname, region = NULL, match = NULL,
  season = NULL)
}
\arguments{
\item{pubgNickname}{The PUBG player nickname that you are requesting statistics for.}

\item{region}{The region of interest, example values include "na", "as", etc... For aggregated results across all regions, use "agg"}

\item{match}{The match type of interest: solo, duo, squad}

\item{season}{The season of interest: 2017-pre1, 2017-pre2, etc...}
}
\description{
This function allows you to pull filtered statistics for a particular PUBG player. Use this function to optionally specify
a particular region, match-mode, and/or season to help filter your query results. If all three filters are applied, the
returning object will be a data-frame with only the results for a particular region, match, and season. See the examples below
for valid filter values.
}
\examples{
fetch_filtered_pubg_player_stats("lazyjustin", region = "na")
fetch_filtered_pubg_player_stats("lazyjustin", region = "agg", match = "duo")
fetch_filtered_pubg_player_stats("lazyjustin", region = "na", match = "duo", season = "2017-pre2")
}
\keyword{filter,}
\keyword{player,}
\keyword{stats}
