% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-pubg.R
\name{fetch_pubg_player_stats}
\alias{fetch_pubg_player_stats}
\title{Fetch Player Statistics}
\usage{
fetch_pubg_player_stats(pubgNickname)
}
\arguments{
\item{pubgNickname}{The PUBG player nickname that you are requesting statistics for.}
}
\description{
This function allows you to fetch player statistics for a single PUBG player. The player nickname is passed in as a parameter.
}
\examples{
fetch_pubg_player_stats("lazyjustin")
}
\keyword{api,}
\keyword{player}
\keyword{stats,}
