% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-pubg.R
\name{find_pubg_player_by_steam_id_64}
\alias{find_pubg_player_by_steam_id_64}
\title{Search for player by Steam ID (64-bit)}
\usage{
find_pubg_player_by_steam_id_64(steam_id)
}
\arguments{
\item{steam_id}{The 64-bit Steam ID that you want to search for.}
}
\description{
This function allows you to search for PUBG players by 64-bit Steam ID. The object returned contains metadata about the player.
}
\examples{
find_pubg_player_by_steam_id_64("10101010101010101")
}
\keyword{api,}
\keyword{player,}
\keyword{search}
