% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-pubg.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key}
\usage{
set_api_key(apiKey)
}
\arguments{
\item{apiKey}{Your public API key.}
}
\description{
This function allows you to set your pubgtracker API key. It should be called prior to making any API calls.
}
\examples{
set_api_key("00000000-0000-0000-0000-000000000000")
}
\keyword{api,}
\keyword{key}
